# docs/developers/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/archive": {
      "post": {
        "operationId": "ProductsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/unarchive": {
      "post": {
        "operationId": "ProductsController_unarchive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/suppliers": {
      "post": {
        "operationId": "ProductsController_addSupplier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSupplierDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/suppliers/{vendorId}": {
      "delete": {
        "operationId": "ProductsController_removeSupplier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vendorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/uoms": {
      "post": {
        "operationId": "ProductsController_addUom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/uoms/{uomId}": {
      "delete": {
        "operationId": "ProductsController_removeUom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/default-bins": {
      "post": {
        "operationId": "ProductsController_linkDefaultBin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkBinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/default-bins/{binLinkId}": {
      "delete": {
        "operationId": "ProductsController_removeDefaultBin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "binLinkId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/components": {
      "get": {
        "operationId": "ProductsController_getComponents",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      },
      "post": {
        "operationId": "ProductsController_addComponent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/components/{componentId}": {
      "patch": {
        "operationId": "ProductsController_updateComponent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "componentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_removeComponent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "componentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
...
    "/webhooks/{id}": {
      "put": {
        "operationId": "WebhooksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ]
      },
      "delete": {
        "operationId": "WebhooksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "ModBM API",
    "description": "Core Forgeron API System endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "productNumber": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "productType": {
            "type": "object"
          },
          "structureType": {
            "type": "object"
          },
          "barcode": {
            "type": "string"
          },
          "listPrice": {
            "type": "string"
          },
          "standardCost": {
            "type": "string"
          },
          "tradePrice": {
            "type": "string"
          },
          "priceLevel3": {
            "type": "string"
          },
          "priceLevel4": {
            "type": "string"
          },
          "purchaseTaxCategoryId": {
            "type": "string",
            "format": "uuid"
          },
          "salesTaxCategoryId": {
            "type": "string",
            "format": "uuid"
          },
          "alternateProductNumber": {
            "type": "string"
          },
          "productGroupId": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          },
          "stateCode": {
            "type": "string"
          }
        },
        "required": [
          "productNumber",
          "name"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "productNumber": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "productType": {
            "type": "object"
          },
          "structureType": {
            "type": "object"
          },
          "barcode": {
            "type": "string"
          },
          "listPrice": {
            "type": "string"
          },
          "standardCost": {
            "type": "string"
          },
          "tradePrice": {
            "type": "string"
          },
          "priceLevel3": {
            "type": "string"
          },
          "priceLevel4": {
            "type": "string"
          },
          "purchaseTaxCategoryId": {
            "type": "string",
            "format": "uuid"
          },
          "salesTaxCategoryId": {
            "type": "string",
            "format": "uuid"
          },
          "alternateProductNumber": {
            "type": "string"
          },
          "productGroupId": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          },
          "stateCode": {
            "type": "string"
          },
          "baseUom": {
            "type": "string"
          },
          "defaultSalesUomId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "defaultPurchaseUomId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          }
        }
      },
      "AddSupplierDto": {
        "type": "object",
        "properties": {
          "vendorId": {
            "type": "string",
            "format": "uuid"
          },
          "supplierPartNumber": {
            "type": "string"
          },
          "costPrice": {
            "type": "number",
            "minimum": 0
          },
          "effectiveFrom": {
            "type": "string"
          },
          "effectiveTo": {
            "type": "string"
          }
        },
        "required": [
          "vendorId"
        ]
      },
      "LinkBinDto": {
        "type": "object",
        "properties": {
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "binId": {
            "type": "string",
            "format": "uuid"
          },
          "isPrimaryPerLocation": {
            "type": "boolean"
          },
          "minQuantity": {
            "type": "string"
          },
          "maxQuantity": {
            "type": "string"
          }
        },
        "required": [
          "locationId",
          "binId"
        ]
      },
...
    }
  }
}